tree::erase method with iterator position.
Erase tree elements specified by iterator pair range. This is similar to
erase signatures for other containers in STL. Some semantics have to be
nailed down but documentation covers it almost completely.

A small test is added to test erase. Inspection of elements is still not
done on the tree test.

Debug build flags are added to enable gdb segfault inspections.