raw_tree foundation class as replacement for Node.
raw_tree is a node replacement class. This is to be the underpinnings of
our smart container. Adding this class ensures that a 'raw' tree is
available to consumers. It also enables the smart tree to program
against it's interface allowing us to switch from a pointer and node
based tree to a contiguous memory store binary tree if necessary.

Interface for the raw_tree is sketched out. Tests are added through a
new test listed in test/