raw_tree_iterator is templated with traversal order.
Currently the traversal order is set in the construction of the iterator
object. This is bad because this is copied around between iterator
objects pointlessly.

Instead we follow a structure similar to how the forward and reverse
iterators for standard sequential containers are built. In this scheme
the iteration direction changes the 'type' of the iterator. This results
in more efficient (in memory and CPU) code as the order drives the code
built and linked.