raw_tree use value templates and left, right enum
Currently the raw_tree class template has methods like 'detach_left' and
'detach_right'. This is not DRY and induces non-DRY-ness in client code.
This will especially be a problem when we try to write the iterator code
for reverse iteration among other things.

This change switches to using an enum class for left and right so that
detach_left becomes detach<side::left> and similarly for other methods.
Call sites now can be dry if they have the side enum.