raw_tree std::array for children; enum converter.
Currently we use C-style arrays for the children of a node. This is fine
but the c++ std::array comes with slightly better access semantics and
type safety. This change switches to using this.

Also an enum converter for 'side' enum class is added to make it more
compact to access the integer value (which is what is used to access
array elements.