raw_tree short names for iterator construction.
This is the multi-pronged analogue to begin-end/rbegin-rend. Under this
scheme we have inlbegin which is inorder-left-begin and prerend which is
preorder-right-end. These are just tiny (inlined) wrappers around the
canonical template member function variants.

Tests are added for testing the various properties of trees using the
various iteration strategies and some other richer modifications.