Test switch to MACROS instead of functions.
In HEAD~3 we added more structure to the tests by introducing functions
to DRY the repetitive testing done for various tree constructions. This
is fine from DRY but a problem from test failure reporting. The reports
report the line number from the file of the function and we don't know
where in the test source file the error occurred.

This change converts the c++ functions to macros so that the actual
testing macros are in-place of the call-site. This makes the test
failure reports report the correct lines.