constexpr static integer variables to avoid verbosity.
Currently to select the children arrays specific index we have to do a
conversion from enum class to integer. This is wasteful (verbose) though
not compute wasteful because this is a constexpr. This change adds a
couple of "hidden" static variables to have the same effect but is more
compact without compromising readability.