iterator allow iterating from past the end.
End iterators are unmodifiable so far. This is fine but if we want to
implement bidirectional iterators we would need iterators which can be
decremented from the end.

This change effects this. Currently we can only increment but because of
the way prefix/postfix and inorder left/right are related increment
forward for one class would be equivalent to decrementing past the end
for another class.

This is a continuing change to adding depth_ to the iterator object.