Accessor is a "smart" pointer without iteration.
Described in a note (which is deleted in this commit), accessor is a
smart pointer that allows us to navigate the length and breadth (and
depth? hah!) of the tree. This accessor doesn't tie up with concepts of
operator++ to be an iterator of any sort, but it doesn't implement a
templated 'next' which takes order and side. This more elementary smart
pointer into the tree is the base for the smarter iterator classes which
allow STL style access to the tree container.

This implementation refactoring also switches to using more public
interface of the tree, this would ensure that we don't tie ourselves
down to the pointer based heap allocated tree node implementation of
trees (the contiguous array model would also fit well).