Iterator is made bidirectional.
Currently iterator is fordward iterator. After the accessor decoupling
of preorder, inorder and postorder next element seek, it becomes much
simpler to effect operator-- on the iterator in a DRY way. This is done
in this commit.

Tests now use reversible iterators instead of adding to a vector and
testing the reverse iteration over that.