Accessor::unsafe_up for efficiency.
Currently we have 'up' which does a couple of tests and then moves the
pointer upward. This is inefficient when calling up from inside the
*increment methods. That's because the context very often implies that
depth_ > 0. These tests done inside the 'up' method are redundant in
such situations.

This commit splits 'up' into 'up' and 'unsafe_up'. The latter is a
private method which can be used internal to the class.