Accessor: Add operator-> to access tree methods.
Currently the accessor allows navigation and dereferencing the *value*
at the node being referred. This is incomplete in that clients would
also like to be able to access the underlying tree for potential
modifications (detaching, attaching etc.).

This change adds operator-> which returns a pointer to the underlying
tree so client sites can seamlessly call tree methods on it.