accessor: common_ancestor, node methods.
common_ancestor is a useful method for tree operations. We define this
on a pair of accessors in this commit.

node is a method to access the actual raw_tree node object. This can be
handy when client code needs to actually the node for using as
arguments. Dereferencing would cause the value held at that node to be
returned.