Single include file.
For header only (and most STL) libraries there is a single file which
includes all relevant fragments and client code uses the specific part
of the fragment it most cares about.

This change also moves in that direction. Instead of a nested dependency
and having client code (like test files) depending on specific header
files, the client code now would be expected to include a single header
file which pulls in all the code and then the client code can use the
specific relevant bits that it wants.