raw_tree: reshape which subsumes rotations.
Currently we only have rotate methods. This is fine but a more general
class of mutations at a node involves two non-rotation reshaping. These
are all generally needed to ensure all possible tree shapes starting
with a line.

This commit introduces reshape with documentation and tests. rotate is
redefined in terms of reshape.