raw_tree: Add const overloads to allow const correctness.
Currently we have some bare minimum methods implemented on the class.
This is fine but is limiting because some of the methods which are
needed for const correctness in other places (const_iterator for
example) are not possible or require complex workarounds for it.

This commit adds some const overloads for methods where it makes sense.