accessor: Setup for different types.
Currently we have a single monolithic accessor. This is fine but we can
get slightly more bang for buck if we have accessors of different types.
The base accessor type is going to have operations for moving around a
tree and performing operations. A derived accessor would allow all the
features that become available when the 'depth' field is present in the
accessor viz. --end(), virtual trees (sub tree), some slightly faster
operations.

The accessor is templated by container type. This is to enable the
accessor to work in const and non-const scopes. In const scopes the
const accessor would allow some limited operations which const liberal
(but correct) operations would enable in those respective scopes.

In this tweaking a few simplifications are effected. Root in raw_tree
now points to self and that's the distinction between it and other
models. This allows for safer operations but also tighter code since we
don't have to do a bunch of nullptr checks in a few places. Conversely
this is also slightly faster. For example up() is now identical to
unsafe_up() (at least with the depth field) in the future this might
change but would still be faster.