traverser: Static functionality for pre/in/post order traversals.
Currently the ordered traversal functionality is baked into the
accessor. This is a problem if we are trying to have different accessors
and different iterators bind to them. Such accessors and iterators
primarily differ in their ability to --end(). This ability difference
makes the cheaper (on memory) accessor/iterator families only forward
iterators. The second are bidirectional.

This change moves the iteration functionality out from the accessor into
the traverser class and it's static function 'next'. The accessors are
reduced to being interfaces which implement the basic ability to
navigate through various relations of a node.