virtual_accessor: Separate virtual subtree navigation.
This class introduces a new accessor 'virtual_accessor'. Earlier
'accessor' used to provide all the goodies of 'virtual_accessor' but
with the need to splice and split trees to handle predicated 'merge'
operations correctly, we need a 'raw' accessor which operates separately
from the 'virtual_accessors'. This is because slicing and splitting will
change the 'depth_' field of the accessor and there is no way to update
it live.

virtual_accessor takes on this field and the roles that come from this.
'accessor' becomes a bare pointer but one which can still enable
'ForwardIterator' criteria with no loss of efficiency in iterators
implemented using this. 'virtual_accessors' would enable bidirectional
iterators for those iterators implemented using 'virtual_accessors'.

Other than this primary change enabling changes are made,
raw_tree_iterator now derives from 'virtual_accessor's to continue to be
bidirectional.