virtual_accessor: Automatic conversion from accessor.
Currently the virtual_accessor has to be constructed with the pointer
and depth fields or using other copy/move constructors. This commit adds
a constructor which converts a regular accessor to the virtual accessor
using 'depth()' function call to compute the depth_ field.

This is effectively converting the regular accessor to the virtual
accessor over the entire tree (not a subtree).