raw_tree: When reshaping, work on fewer nodes.
In certain reshape operations, the result can be valid if some of the
nodes are not present. For example when doing R-R reshape and the right
child of the current node is not present, we can still do the reshape.
The resulting two node tree will simply be right linked.

This change enables this. It would reduce the number of client sites
would have to check and handle.