accessor: Remove root check from 'up'.
Currently we have an 'up' method publicly visible and an 'unsafe_up'
which is used internall by the object. This separation is not needed as
all client sites can effect the 'safe' up with a test. In practice the
call-site algorithms almost always are checking for 'root-ness' before
deciding to move up.