accessor: Remove is_const template argument. 'ancestor' method.
is_const is an essential feature on the accessor to maintain const
correctness ala const_iterator. This feature however is implemented
incorrectly by specifying it as a template boolean value argument that
client sites can set.

This may have been correct when the first template argument was 'T' the
value type the accessor handles but since we moved to 'Container' type
which the accessor refers to this has become superfluous. We can compute
the const-ness from the first template argument. Also the client cannot
be expected to set this correctly.

This change removes this argument and derives it from the first
argument.

This change also adds a new method 'ancestor<side>' which moves to the
closest ancestor of a node on a particular side.