raw_tree_iterator: Move out of raw_tree.
The iterator class has never been dependant on internals of the raw tree
object or types. The iterator is completely dependant on
(virtual?)accessor and the directions of iteration only.

To effect const_iterator the way we have started off in 'accessor' (as
the const qualifier of the template container argument) we need to
derive from the accessor in the same vein. This is at least a little
confusing if we have the iterator derive from const. For example:
raw_tree<T>::iterator : public accessor<const raw_tree<T>> ...
is weird. Also there should not be a way to specify the const qualifier
on the nesting class (hah!).

This change moves the iterator out from nesting. Related call-site and
definition changes are made.

test is updated for more comprehensiveness but also carries some of the
updates related to the accessor and iterator structure changes.