predicated_tree: A tree wrapper which upholds predicates.
predicated_tree is a 'smart' container. predicated_tree is to raw_tree
in a similar way that set is to list. Full notes are in the header file.

predicated_tree needs a couple of companion classes, comparator which
describes properties which the tree has to uphold. mutator which is a
'free' object which manipulates tree elements.

Tests are added for testing basic operations. One benchmark test is also
added.