docs: Project readme. Remove doxygen autobuild.
Moved the index.md file to //README.md. This is better for github
project homepage.

Removed doc/CMakeLists.txt as we are not able to successfully auto build
documentation from source. We will instead auto-build documentation on
local machine and publish it using github-pages.

Made a small change in the handling of version information in the cmake
file.