predicated_tree: make_* helper fns; default template args
Currently comparator and predicated_tree both require full specification
of template arguments. This results in a fair bit of boilertplate code
for the predicated tree (where both type has to be specified and the
construction argument has to be specified).

This change switches to using template and constructor argument defaults
liberally to enable client code to initialize objects in a more
streamlined way (more consistent with STL containers).

Additionally usable defaults are specified for the comparator (height is
indifferent and left is std:less when possible). This would make the
client code work more seamlessly, especially if the contained-type has a
usable operator< defined over it.

make_* helper methods are added to allow client to only specify one set
of arguments (template or function arguments) and allow deduction for
the rest of the setup.