impl subdirs for *.hh files. relative-path includes.
Currently we store the *.hh files parallel with the corresponding *.h
files. This is a little noisy on the filesystem especially when a reader
is trying to grok the interface of the library.

This change switches to using an 'impl' subdirectory wherever necessary
(inside include/ of course) for storing the implementation of template
interfaces. The *.hh files are moved inside such directories.

A slightly unrelated but essential enabling change (essential for
portable installing and namespacing) is that all library internal
'includes' are now relative paths. This ensures that transplanting the
include directory inside a package would not break references for
imports.