mutator: sift_down for erase and make_heap; remove zip.
Currently make_heap is incorrect. Currently re-heap is done over a local
neighbourhood. This is not correct because the re-heap might cause
locality to change and the heap predicate needs to be re-evaluated in
the new neighbourhood.

This change introduces a new method sift_down which does the traditional
heap sift_down operation. This operation is used by the make_heap
methods to correctly do heaping of a sorted tree.

A variant of this method is also introduced sift_out which is used by
erase. Erase is strictly speaking correct but using this analogous
method ensures we don't introduce unnecessary algorithms where not
necessary.

The zip method(s) were present to allow joining two trees which were
sorted in L predicate. With the heap down operation, we can have the
same effect and erase (sole consumer of zip) doesn't need to use that
anymore. So we remove the zip methods.