accessor: Allow non-explicit instantiation.
Implementation code and client code call sites can be DRY with this.
Almost always when an operation is being attempted on a subtree or a
tree if the implementation only supports the tree structure as argument
then we can work but would require call site to sometimes convert an
accessor to a tree. Converse is also true. To help the call site, the
implementation sites would have to implement multiple variations (ones
accepting the tree and ones accepting the accessor).

This change allows the accessor to non-explicitly be constructed at the
root of a tree. This would allow implementation sites to only implement
the accessor version and all possible call site options would work
correctly.