raw_tree: Move operations don't update parent of moving node.
Currently move operations, update the parent of the target node as well.
This has the effect of vacating the tree where a node resided and moving
it to a different tree altogether.

This change switches the move semantics to move a node *into* a tree.
This is probably more useful and also probably more in line with the
common-use semantics of move across other object and container types.
Move operation still detach nodes from source node.