ptree: Move mutator methods into ptree except tree transformations
Currently all tree operations (find, insert, erase etc.) are done
through the ptree by the mutator. The ptree is just a proxy which binds
the mutator to the raw tree. This is not necessary as we move to a
better structure:
- Ptree is no longer a proxy. It effects all tree methods which perform
  operations on a ptree itself.
- Mutator retains (and after renaming will assume) all tree
  *transformation* methods i.e. functions where the predicates for trees
  are changing.

Comparator might go away and the predicates will be passed in raw. Some
methods would need a better place (like sift down etc.) which are usable
inside transformations but also tree operations.