comparator: Remove vertical/horizontal. Move bits to ptree.
Currently comparator is a composite predicate handler. This has become
redundant since we directly can use left and tall call signatures in
almost all places. The exception is when we need to do templated call.

The code is changed to simplify the comparator (as an incremental step
toward removing the comparator altogether). The relevant client
call-site functionality is updated. horizontal is moved to ptree for the
clip function.

This and removing the comparator in general should make the code leaner
and easier to reason about.