comparator: Remove. Inline functionality.
Comparator was an unnecessary component. It seemed like it was necessary
but with more correct and semantically consistent interfaces on the
various methods we can do away with this entirely. Call sites now
directly use one or the other predicate. Related functionality is
inlined where necessary. This is in continuation with the set of changes
necessary for better code structure (no mutator, no comparator but a
tree transformer).

Client code (where predicated tree or mutator type information is used
is updated to reflect new layout.