predicate:gardener class for tree transformations.
Currently we can build trees from scratch, perform some tests and
*subsume* trees by assuming them to be correct. This is fine but there
is another feature which is essential. Given a predicate-pair converting
an existing tree to another tree with a different (possibly partially
overlapping) predicate pair.

This is needed for example if a tree is built with indifferent height
predicate but needs to be rebuilt with a different specific predicate.
This would only affect heaping but not sorting order.

This commit introduces a class template for doing exactly this. The
interface is generic enough to allow all optimization opportunities and
there are two implementations (a basic and a stable heap one).