predicated:tree static assert for predicate type matching.
Currently we don't have a static assert on matching of predicate's
callable argument types and the types of the tree nodes. This leads to
long list of compile errors when making a mistake. To make the library
easier to use for developers, it would be nice to have this be
static_asserted so that a useful error message can be printed for the
client code.

It is also a nice indirent mechanism for documentation of the codebase.