predicated:tree use `taller` methods if present.
Just like we use std::less if operator< is defined on the tree data type
T, this commit adds the feature to use a `taller` method if it is
defined on the data type. An additional simple wrapper (similar to
std::less) is created for using this method.

Minor class renaming to make it more std-name-like.