predicated:heap_iterator is used for iterating elements heap.
Currently we have in, post and preorder iterators on the tree. These are
iterators available on the raw tree.

This commit introduces the heap-iterator iterator which consumes the
tree to generate the sequence of elements in `tall` order.  The
heap-iterator is unique to the predicated tree, since heap iteration
requires popping and then sifting-up through the values in the tree
which involves doing comparisons through the predicate.

A small test is added for testing this iterator.