predicated:wrapper macros for easier construction.
Currently the wrapper class template requires specifying derivable type
information. This is problematic for the call site because that is
expected to setup matching parameters. This commit adds a couple of
macros that make the setup DRY and composable.

Some of the current call sites where wrapper objects are used are
updated.