docs: Inline manual, predicates-group, remove index.md
Currently we have an index.md which is empty and used as the root
index.html file in the generated documentation. This is quite broken
because we want to mainpage to be the manual.

This change fixes this by making the README.md as the mainpage for the
generated documentation.

Generated group for 'basic', 'predicated' and 'predicates'.

Fixed documentation in a few small places.