predicate:tree more tests (types) and fix in a test.
Currently the insert predicate tree test has a bug which uses the same
type for the heap and less operations. This means instead of 16
different combinations of tests, we run only 4 tests 4 times each. This
bug is fixed in the commit.

The heap conversion test is also expanded in scope to operate over a
range of types using the same mechanism (TEMPLATE_PRODUCT_TEST_CASE).

Some missing includes in a couple of the newer modules are also fixed.