raw:s/is_side/get_side/ to get side instead of querying.
The non-templated is_side method on raw_tree is not useful and using the
templated version on some client sites does not bring efficiency which
the compiler can forward anyway. This change switches the interface to
return the side via a `get_side` call which client site can pass forward
directly into a mutation function (non-templated detach for example).