raw:detangle headers
Currently header includes are a mess. It works because in most binaries
which we have to compile we are directly including tree.h which has a
correct sequence. This is not correct though as a reorder of includes
would cause compilation errors.

This commit moves side and traversal_order into a separate enums.h*
header library. Includes for these files is done correctly in places of
use.

All other headers which were including tree.h directly are now fixed to
include the actual required library elements.

This can be a useful first step in separating the library into two
independent libraries (raw_tree and predicated_tree).