predicate: Move raw tree code to separate repo.
Currently we have a single repository for both the raw tree and the
predicated tree containers. This is not ideal since each can develop on
it's own.

This commit removes the raw tree code out from this repository. This is
now checked into a separate repository from which we can clone and
include as we please.

Documentation is updated to reflect the specifics of this library.